move workspace code into subdirectory
