+++ src/lib/workspace/EDASession.tsx
